Refactors ScoreId to FacilityId in DTOs/Entities
Renames ScoreId to FacilityId in GroundwaterScore and OnSiteScore DTOs and entities to align with the intended relationship.

Updates Parcel DTOs and entities to use FacilityId instead of LocationId to properly associate parcels with facilities.

Updates repositories to use FacilityId for retrieving GroundwaterScore and OnsiteScore.

Includes necessary navigation properties to facility entity.

Updates seed data and FacilityDetails to reflect changes.