Makes ChemicalId nullable in score entities
Makes the ChemicalId property nullable in GroundwaterScore and OnSiteScore entities and their DTOs. This allows for scenarios where a chemical is not always required.

Adds constructors to entities (HsrpFacilityProperties, Location, OnSiteScore, GroundwaterScore, Score, Status) which take only the FacilityId parameter.
These constructors initialize the Id with a new Guid and set the FacilityId.

When a new HSI Facility is created, placeholder objects are created for HsrpFacilityProperties, LocationDetails, ScoreDetails, GroundwaterScoreDetails, OnsiteScoreDetails, and StatusDetails. This ensures that these related entities are readily available.

Removes the unnecessary "StringOrNone" format specifier from the textarea displays for comments in the Facility Details page.