Improves Facility Details page functionality.
Adds functionality to delete substances, view a summary of substances.

Adds a map link to Facility Details page based on latitude and longitude coordinates.

Adds ActiveTab to the temporary data.
This allows the user to be redirect back to same tab they where on after
editing a section of the facility details.

Filters "HWTF Master Project" events on the Event Add/Edit pages.