Adds Final RC and RQ properties to Chemical
Adds new properties 'FinalRc' and 'RQ' to the Chemical entity and related DTOs.

These properties are added to the Chemical entity, as well as the ChemicalCreateDto, ChemicalEditDto, and ChemicalSummaryDto.
The ChemicalRepository is updated to handle these new properties.
The Add and Edit pages for Chemicals are also updated to include input fields for the new properties.
The Index page is updated to display the new properties.
Seed data is updated to include values for the new properties.