Updates display names for clarity
Updates the Display names for better readability and clarity in the UI.
Specifically, renames "Description" to "GW Comment" for GroundwaterScore DTOs and Entities, and to "On-Site Comment" for OnSiteScore DTOs and Entities.
Also, renames "Use alt language for Site Summary Report" to "Use alternate language from comments field" for Score DTOs and Entities.
Adds missing display names for status fields.