Adds descriptions to item lists
Adds a description field to Funding Source, Groundwater Status, Overall Status, Soil Status, and Source Status entities and DTOs.

Updates the item list repositories to include the description in the display name.

This change enhances the usability of dropdown lists by providing more context for each item.

The DisplayName property is added to the DTOs and entities for better readability.

Fixes #807