Refactors facility form validation for clarity
Improves code readability and maintainability in the facility form validation logic.

Introduces an enum for validation error messages and centralizes error message generation. This reduces code duplication and makes it easier to manage and update validation messages in the future.

Updates javascript on add and edit forms to improve user experience.