Enables chemical name autofill
Changes the chemical search API to accept the query as a query parameter instead of a route parameter.

Updates the JavaScript in the add and edit substance pages to properly set the chemical information from the search results. The previous implementation attempted to access the first element of individual properties which were already the desired string values.

Fixes #815