Modifies Score and Substance tables
Updates the database schema to better support scoring
functionality.

Removes the ScoredById column from the Scores table,
as this relationship is not needed.

Renames UseForScoring to UseForSoilScoring in the
Substances table for clarity.

Adds UseForGroundwaterScoring column to the Substances table
to allow substances to be scored for groundwater impact.

Fixes #858