Enhances phone number formatting and validation
Improves the phone number format on contacts by adding a regex expression for validation in both the create and edit phone DTOs.
Also, updates the labels of data annotations to improve readability.
Corrects the sorting of events on the facility details page and
updates the names of display labels for Groundwater and Onsite scores.
Adds libphonenumber-csharp package.