Enhances Event data displays and folder links
Improves Event-related data displays by adding display names and data type annotations to DTOs and entities.
This change also enhances the presentation of Event data in the UI, specifically on the Add, Edit, and Details pages.

Adds a link to the "Master Project Complaint Folder" on the Facility Details page for facilities marked as "COMPLAINT".
This provides users with direct access to relevant complaints documentation.